+++ lib/view_models/provider/tap_bar_provider.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class TapBarProvider extends ChangeNotifier {
+  int _selectedIndex = 0;
+
+  int get selectedIndex => _selectedIndex;
+
+  void onChangeIndex(int index) {
+    Future<void>.delayed(const Duration(milliseconds: 500), () {
+      _selectedIndex = index;
+      debugPrint('TapBarProvider | Current Index: $_selectedIndex');
+      notifyListeners();
+    });
+  }
+}