+++ CHANGELOG.md
@@ -1,3 +1,15 @@
-## 0.0.1
 
-* TODO: Describe initial release.
+# Changelog
+
+## 1.0.0
+
+- Initial release of `reusable_tab_bar`.
+- Provides:
+    - `TapBarProvider` for managing tab selection with `ChangeNotifier`.
+    - `TabBarCubit` and `TabBarState` for managing tab selection using `Bloc`.
+    - `AppTabBar` widget for creating customizable TabBar layouts.
+    - `AppTabBarWithSelectorButton` widget for custom tab navigation using external selector buttons.
+- Supports features like:
+    - Custom app bars, floating action buttons, drawers.
+    - Dynamic tab selection and state management.
+    - Full flexibility for styling and behavior control.